+++ featurebyte/service/target.py
@@ -3,37 +3,19 @@ Target class
 """
 from __future__ import annotations
 
-from typing import Any, Callable, List, Optional, Union
-
-import time
+from typing import Any, Callable, Optional, Union
 
 import pandas as pd
 from bson import ObjectId
 
 from featurebyte.common.model_util import parse_duration_string
-from featurebyte.common.progress import get_ranged_progress_callback
 from featurebyte.exception import DocumentCreationError, DocumentNotFoundError
-from featurebyte.logging import get_logger
 from featurebyte.models.feature_namespace import DefaultVersionMode
-from featurebyte.models.feature_store import FeatureStoreModel
 from featurebyte.models.observation_table import ObservationTableModel
-from featurebyte.models.parent_serving import ParentServingPreparation
 from featurebyte.models.target import TargetModel
 from featurebyte.models.target_namespace import TargetNamespaceModel
 from featurebyte.persistent import Persistent
-from featurebyte.query_graph.graph import QueryGraph
-from featurebyte.query_graph.node import Node
 from featurebyte.query_graph.node.schema import TableDetails
-from featurebyte.query_graph.sql.common import REQUEST_TABLE_NAME
-from featurebyte.query_graph.sql.feature_historical import (
-    NUM_FEATURES_PER_QUERY,
-    PROGRESS_MESSAGE_COMPUTING_FEATURES,
-    TILE_COMPUTE_PROGRESS_MAX_PERCENT,
-    get_feature_names,
-    get_historical_features_query_set,
-    get_internal_observation_set,
-    validate_request_schema,
-)
 from featurebyte.schema.target import ComputeTargetRequest, TargetCreate
 from featurebyte.schema.target_namespace import TargetNamespaceCreate, TargetNamespaceServiceUpdate
 from featurebyte.service.base_namespace_service import BaseNamespaceService
@@ -44,91 +26,8 @@ from featurebyte.service.namespace_handler import (
     validate_version_and_namespace_consistency,
 )
 from featurebyte.service.session_manager import SessionManagerService
+from featurebyte.service.target_helper.compute_target import get_targets
 from featurebyte.service.target_namespace import TargetNamespaceService
-from featurebyte.session.base import BaseSession
-
-logger = get_logger(__name__)
-
-
-async def get_targets(  # pylint: disable=too-many-locals, too-many-arguments
-    session: BaseSession,
-    graph: QueryGraph,
-    nodes: List[Node],
-    observation_set: Union[pd.DataFrame, ObservationTableModel],
-    feature_store: FeatureStoreModel,
-    output_table_details: TableDetails,
-    serving_names_mapping: dict[str, str] | None = None,
-    parent_serving_preparation: Optional[ParentServingPreparation] = None,
-    progress_callback: Optional[Callable[[int, str], None]] = None,
-) -> None:
-    """
-    Get targets.
-
-    Parameters
-    ----------
-    session: BaseSession
-        Session to use to make queries
-    graph : QueryGraph
-        Query graph
-    nodes : list[Node]
-        List of query graph node
-    observation_set : Union[pd.DataFrame, ObservationTableModel]
-        Observation set
-    feature_store: FeatureStoreModel
-        Feature store. We need the feature store id and source type information.
-    serving_names_mapping : dict[str, str] | None
-        Optional serving names mapping if the observations set has different serving name columns
-        than those defined in Entities
-    parent_serving_preparation: Optional[ParentServingPreparation]
-        Preparation required for serving parent features
-    output_table_details: TableDetails
-        Output table details to write the results to
-    progress_callback: Optional[Callable[[int, str], None]]
-        Optional progress callback function
-    """
-    tic_ = time.time()
-
-    observation_set = get_internal_observation_set(observation_set)
-
-    # Validate request
-    validate_request_schema(observation_set)
-
-    # use a unique request table name
-    request_id = session.generate_session_unique_id()
-    request_table_name = f"{REQUEST_TABLE_NAME}_{request_id}"
-    request_table_columns = observation_set.columns
-
-    # Execute feature SQL code
-    await observation_set.register_as_request_table(
-        session, request_table_name, add_row_index=len(nodes) > NUM_FEATURES_PER_QUERY
-    )
-
-    if progress_callback:
-        progress_callback(TILE_COMPUTE_PROGRESS_MAX_PERCENT, PROGRESS_MESSAGE_COMPUTING_FEATURES)
-
-    # Generate SQL code that computes the targets
-    historical_feature_query_set = get_historical_features_query_set(
-        graph=graph,
-        nodes=nodes,
-        request_table_columns=request_table_columns,
-        serving_names_mapping=serving_names_mapping,
-        source_type=feature_store.type,
-        output_table_details=output_table_details,
-        output_feature_names=get_feature_names(graph, nodes),
-        request_table_name=request_table_name,
-        parent_serving_preparation=parent_serving_preparation,
-    )
-    await historical_feature_query_set.execute(
-        session,
-        get_ranged_progress_callback(
-            progress_callback,
-            TILE_COMPUTE_PROGRESS_MAX_PERCENT,
-            100,
-        )
-        if progress_callback
-        else None,
-    )
-    logger.debug(f"compute_targets in total took {time.time() - tic_:.2f}s")
 
 
 class TargetService(BaseNamespaceService[TargetModel, TargetCreate]):
